+++ blocks/navigation/block.ts
@@ -1,0 +1,4 @@
+export interface Parameters {
+  backLabel?: string;
+  forwardLabel?: string;
+}