+++ CHANGELOG.md
@@ -9,7 +9,12 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Block(`@amsterdam/navigation`): Add navigation block.
 - CLI: Add support for `@import` in organization style sheets.
+
+### Fixed
+
+- Block(`@appsemble/form`): Fix `maxLength` not being passed to string input fields.
 
 ## [0.8.9] - 2019-10-02
 
@@ -28,7 +33,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.8.7] - 2019-09-16
 
-### Fixes
+### Fixed
 
 - Server: Add missing migration from 0.8.6.